Update OpenStack Operator
The existing OpenStack operator will only work for
Python 2.7 Airflow Containers and will fail to function
properly for Airflow Containers running on Python 3.x

This update will fix the problems with the environment
variables.

The current assumption is that the shipyard.conf will
contain the required information for the OpenStack
environment and it will be consumed by Airflow Dag to
execute the OpenStack commands.  This will likely change
in future when DeckHand is ready for integration as the
required information will be retrieved from DeckHand instead

The xcom portion has been removed as it does not really
serve any purpose at the moment.  It might get added back
in future depending on the needs.

Change-Id: I0c5fb56462fdbc43a20897373118f510cbb2e01c