Add Health Check Endpoint
The current design for health check will be for the UCP
component to return an empty response to show that it is
alive and healthy.

This patch set adds a /heath endpoint to Shipyard

Keystone authentication is not required for this health
check and query can be made by anyone.  Extended health
check which offers more information/details will require
authentication and will be introduced at a later stage.

Change-Id: Ib06355a9e6a97727c1340f371fe8d496bf3861f9