Update deploy_site Dags & Operators
There is a need to update the existing dags, i.e.
deploy_site, redeploy_server as well as update_site
in order to make use of xcom. This is needed in order
for the shipyard ACTION API to trigger the workflow.

A new subdag for drydock workflow has to be created
as well in order to fit into the flow of the deploy
site dag.  DryDock Operator was updated as well so
that we can use it with the deploy_site dag.

The uuid of the ACTION performed from Shipyard will
be logged as well so that we can use it for logs
correlation.

Change-Id: I25f7486a0510431d663547988fb5884a9845d2aa