Update update_site Dag
Update the flow for DryDock.  The flow for update site will be
similar to the one for deploy_site as far as DryDock is concenred
as DryDock will only be used for deploying nodes.

A possible use case where we can use drydock with this dag will
be when we are trying to add nodes to an existing site. The flow
for drydock will be the same as that of deploy_site in such scenario.

Change-Id: Icfb0e80ca926fe3a42e88db75e9dec18f073b4c8