Update Action API documentation
This patchset aims to update the documentation for the action api of
shipyard to match the functionality being developed. Request and
response examples have been included for those known entities. The
representation of validations has not been finalized, and is therefore
missing from this patchset.

Other changes made include adding some notation explanation for the examples,
Duplicating the name of the API to the HTTP verb associated to make for
better readability, and some cleanup of existing verbiage to use more
precise terms.

Change-Id: Ibf80bd1f5782f346bdd628337167b5e64670d978