Add DeckHand Operator
This P.S. will implement part of the requirements for the
DeckHand Operator.  It will add the logic to retrieve the
'committed' revision that will be used during deployment.

The '/validations' endpoint has not been implemented in
DeckHand and will be worked upon shortly.  Another P.S. will
be submitted at a later date to extend the functionalities of
the DeckHand Operator to support 'validate_site_design'.

Change-Id: I82766dcb8ae93b2a8eb72efc6134f135edd03d51