Update deckHand_get_design Dag
This dag is likely going to be used by Shipyard in actions other
than 'deploy_site'.  Hence we should not be hard-coding values in
the dag as it limits its reusability, e.g. it is likely going to
be used in 'update_site' and in 'redeploy_server' as well.  Hence
there is a need to make use of the values that are passed into the
function instead.

This P.S. is meant to correct the current behavior.

Change-Id: Ia88e91dbb114c34530e3240f120df8c11325090b