Implement the inquiry API for Shipyard.
This allows for query and browsing of all items that have been
invoked via Airflow, whether through the Shipyard front door or not.

Processing for this set of endpoints has been placed in a subdirectory
of the control directory. A later refactoring will move the other API
resources and helpers to their own respective directories.

Change-Id: I7242fba80ad745ef5bcd41d7d07a320dfebb1dd7