Revamp DryDock Operator and Dag
There has been changes to the way in which DryDock
YAMLs are consumed by DryDock.  We will make use of
a self-hosted NGINX container on the Genesis Host
to house the YAMLs while waiting for the integration
work to be completed between DryDock and Deckhand.
This is the current approach taken by CI/CD.

This P.S. is meant to align Shipyard/Airflow with the
current implementations in CI/CD Pipeline.

Change-Id: I7136aa74e33f980589820c75159fa3259ee79976