Update 'deploy_site' dag for Integration Testing
We are currently in the process of reviewing and merging
patch sets from various UCP components that are needed for
the 'deploy_site' dag to work properly. In order to proceed
with integration testing with the CI/CD team, there is a need
to rename the 'deploy_site.py' dag as 'deploy_site.wip' while
we sort out, review and merge the outstanding patch sets.

We will only include the DryDock workflow here for our current
testing. Updates will be made to this test dag as we progress
along with the integration testing and UCP code reviews/merge.

The 'deploy_site.wip' dag will be removed once the system is
fully integrated from an End-to-End perspective.

Change-Id: Ic6b1b22c259d358de1c5f65ee6e452e86ed1d74b