Update Shipyard Helm Chart
This patch set removes the shipyard config, policy and paste.ini
template from the existing Shipyard Helm Chart.  This is done to
align with the current approach in OpenStack Helm.

1) Remove shipyard config template
2) Remove shipyard policy.yaml template
3) Remove shipyard api-paste.ini template
4) Update related template files

There has also been a recent change to the Helm Toolkit which will
break the current implementation of the Shipyard Chart

The changes in Helm Toolkit were made to the 'images' definition
in values.yaml to facilitate adding the option to prefix image
name etc

This P.S. will also update the Shipyard Chart to align with the
recent changes in Helm Toolkit

Change-Id: Ie79fd9da2c9a577027dd0dddbcca6b7f7b3b4f6f