Logging to use request scope provided by uwsgi and standard logging filters
This change introduces a logging filter that sniffs for the presence of uwsgi
being provided to the application, and if so, uses it to provide request scoped
logging variables. This change allows for more standard logging pattersn to be
followed throughout the rest of the API code while still providing information
related to the request, such as the request ID and the external context marker
that may be used by a client of the software.

Change-Id: I82f9070e25a97043ddf4660635595c473a38cda2