Rename 'Armada_deploy_site' Dag for Testing
In order to proceed with integration testing with CI/CD,
there is a need to stop the 'armada_deploy_site' dag from being
copied over to the Airflow containers as the Armada client codes
is still being reviewed at the moment.

This change is temporary and will be reversed once the Armada
client related codes are reviewed and merged.

Change-Id: I10dc4bed257ae63133a730e0b6c20d9885d47eed