Increase Query Time for deploy_site script
It takes about 31 minutes to execute the drydock
bare metal deployment in the CI/CD lab. There is
a need to make the query time a variable that can
be changed according to needs. This will help to
reduce the volume of query response in logs.

The default will be set to 90 seconds. User can
change it by specifying the value of $1, i.e.

$ tools/deploy_site.sh 180

There are some problems with the print outs of the
final json results which is corrected here as well

Change-Id: I7c6e2c5b3a23b4daf36cf78cec693899ba2cab5c