Update Armada Operator
As part of integration testing, we will need to be able
to read the Armada YAML file from remote server as the
integration work between Armada and DeckHand is still
ongoing at the moment.

This P.S. updates the logic in which the Armada Operator
retrieves the Armada YAML

We are aligning the directory and file names with what is
in CI/CD. This will require minor changes to the DryDock
Operator as well.

Change-Id: I26c7c1dc3628662f34fbf7cd5ef916cf34f08243