Fix typo in Shipyard Chart
There was a need to make changes to the variable names when we
merged the Airflow and Shipyard charts. This resulted in typo
in variable name in the service-shipyard-ingress.yaml which stops
the service from being deployed.

This P.S. is meant to correct that behavior

We also need the public endpoint for shipyard to be on port 80
so that the CLI will work properly with the Ingress Controller

Change-Id: I0483e8ab9e3eb7839149413311abb8c1475f59fa