Update deploy_site script
It seems that there is a possibility for Airflow to respond
with 'unknown' state for some corner cases. This will cause
the deploy_site script to get into infinite loop. The codes
in Shipyard has been updated to return 'Unknown (*)' as the
state of the action lifecycle in such situations.

This P.S. is meant to update the script and to set a default time
out of 1.5 hr for the workflow. The time out can be changed by
setting a value for $2 (Note that we will need to set query_time,
i.e. $1 as well)

    $ ./deploy_site.sh 120 7200

Change-Id: Ic56ce85f55e9b5fa0c3f537c6158ddc3dc56fab2