Update DryDock Operator
There is a need to update the DryDock Operator as DryDock
has been integrated with DeckHand. The YAMLs should be
retrieved from DeckHand rather than from nginx server.

This P.S. updates the DryDock Operator to account for the
recent changes. It also reintroduce the workflow for DeckHand
in deploy_site dag.

Change-Id: I4f00ad198da1cbab426394d56f7eb90ef260e0a1