Update DryDock Operator
It seems that there is a newly introduced bug in DryDock where
the result of a completed task is being reported as 'incomplete'.
This P.S is meant to update the DryDock Operator.

The DryDock Operator will be updated again when the bug is fixed in
DryDock.

The bug in DryDock is currently tracked by [0]

We will also add additional logic to check that all the nodes have
joined the cluster before declaring the 'deploy_node' task as completed

[0]: https://github.com/att-comdev/drydock/issues/73

Change-Id: I661995621520fa3d621623da41d984159e251edb