Update Shipyard Operators
There are a couple changes in this Patch Set:

1) There has been recent changes in the genesis process and
the Tiller pod now resides in the ucp namespace instead of
kube-system. As such it no longer takes on the IP of the Genesis
Node. The YAMLs continue to be hosted on the nginx server
which resides on the genesis node. Hence we will need to get
the IP of mass-rack pod instead in order to resolve to the
correct IP of the nginx server.

Note that this logic will change in near future when Armada
is integrated with DeckHand. The nginx server concept will
go away at that time.

2) Bug fix in 'cluster_join_check_backoff_time' variable type
in DryDock Operator. That variable needs to take on integer value.

3) Updates to the 'get_k8s_pod_port_ip' Operator. Note that it is
possible for pod to have an IP with no port. For instance the
maas-rack pod takes on genesis node IP and has no port associated
with it. We will assign the value 'None' to the port value in such
cases.

Change-Id: I07dd922c98c63e03a8963f7767b457c932ee772b