Conform expected validation responses to status standard
Upon review of standard responses, the validation responses
were found to be using valid/invalid instead of success/failure.
This change changes the values used to be success/failure for the
status field in validations. Additionally, this also cleans up the
misuse of api version 'v1', making it v1.0 instead on the status
responses.

Change-Id: Ia10674158648ef0d0c42badc4bd19e1b48b255d9