Update deploy_site script
It is possible for the script to run for a prolonged period of
time. This can cause the keystone token to expire and we will
end up with `null` response from Shipyard when we query the
status of the task.

Note that we will retrieve a new token in the above scenario
and use the new token to query the status.

Note also that the 'deploy_time_out' variable should actually be
the number of loops rather than time out (in seconds). It will
take on the default value of 60 loops (based on 90 seconds back
off per cycle, i.e. 60 * 90 = 5400 seconds = 1.5 hrs) if value
was not provided by user during run time.

Change-Id: I83f503222899ace3d04aac303d2790e4b5d22d2d