Update Armada Operator
This P.S. will add 'validate site design' workflow to the
Armada Operator. We will make use of the /validatedesign
endpoint in Armada for this purpose.

We will also make use of deckhand reference url instead of
nginx server to retrieve the YAMLs and will remove the existing
'armada_validate' step as that would be done during the initial
stage where the site design is validated.

The reference to the genesis node in the Drydock operator will
be removed as it is no longer needed.

Change-Id: I63e9bcd44d814d12db866032dd56fdfdc28d7b4d