Add Wrapper Script for Shipyard CLI
Installing the Shipyard CLI on the Genesis host will introduce
large number of packages on the host machine and is not the
desired approach. Hence we will want to use a docker container
to execute Shipyard CLI whenever possible. However, it is
troublesome to pass all the required parameters everytime
when we execute the Shipyard CLI with docker command.

This patch set attempts to create a thin wrapper script to make
it easier for user.

Change-Id: I06235399cc7c9a22169f39edf948dc5e573652c5