Shipyard-270 - Rendered document schema validation check
The story requires Shipyard to get the rendered document from
Deckhand before proceeding with the rest of the workflow.

This is to ensure that the schema validation at Deckhand is
good and to fail early if need be.

This p.s. attempts to make use of the deckhand client to get
the list of rendered documents. Note that we will need to install
the DeckHand client libraries in the Airflow image in order for
this to happen.

Change-Id: I22f4165be101edfa4cef9bead6dd3d3765146c6e