Update Scripts
We will no longer need to manually update the '/etc/hosts' file
for 'in-cluster' scenario with the bug fix in [0]. This patch set
is meant to update the scripts in the tools directory to reflect
that change.

Note that user will still need to ensure that the DNS is set up or
'/etc/hosts' file is properly updated if he/she is executing the
workflow from outside the cluster, e.g. via jump server

[0] https://review.gerrithub.io/#/c/398739/

Change-Id: I0580cfc473b0f1b0ed824a578d3e6c89dd6f1443