Accommodate mutiple Armada manifests
For site designs, there is a desire to have two different
manifests representing the bootstrap deployment and full site
deployment. This change provides the hard-coded 'full-site'
reference to the manifest to be used for deployment. Making this
value configurable is forthcoming in future enhancements.

This change will require that an Armada manifest document named
'full-site' would be included in the collection of yaml documents.

This change should not be merged until Armada is also capable of
validating designs that contain multiple manifests.

Change-Id: I526c347d86518029bd3959c0230e2e7f2baa293c