Remove airflow config template
This patch set removes the (pre)generated config ini file from
airflow. The configuration will now be pulled directly from
values.yaml which will be inline with OpenStack-Helm's approach.

This will do away with the need to maintain the verbose .conf.tpl
in the repository as mentioned by Tin in his comments for [0].

[0] https://review.gerrithub.io/#/c/400925/

Change-Id: I5a9766e52536ac9b143b397faa3563e69dfb6bf3