Remove logging_config_class from values.yaml
We are seeing the following error [0] in the Airflow
Web GUI which prevents user from reading the workflow
logs from the GUI.

This is happening as the Airflow Web Pod is not able
to directly access the volume of the Airflow Worker
Pod.

This patch set will remove the parameters that are
causing this behavior and revert back to the default
system configuration which was shown to be working
properly in our local test environment.

[0] Error Message

Task log handler task does not support read logs.

Change-Id: I71cc9ebd5f6571b486af4d77dbd89f234e8dd3b3