Harden 'get_k8s_pod_port_ip' Operator
We note that 'list_pod_for_all_namespaces' can return results
in any order. This means that we can end up with inconsistent
results if we have search string that matches the partial names
of 2 or more pods.

This is especially true when the 'get_k8s_pod_port_ip' Operator
only selects the first pod with the name that contains the
search string. Hence there is a need to update the Operator to
avoid getting into such situation.

We will also require namespace to be passed in to avoid any
ambiguity.

Change-Id: I6b93d85ea25b5a30d8a115d78bfa4a51198d6bcb