Make log levels configurable
Allows logging levels to be specified in the shipyard.conf so that
specific logger levels can be set by logger name.

The default option specified for keystoneauth prevents repetitive
logging of information that is generally not useful and is polluting
the logs with extraneous info.

Removes or reduces the logging level of some information, including
database queries and response messages. Response messages are now
only logged at debug in the case of an error response.

Change-Id: I895439e78c6ba71b4b7d6bd7dd26b768221e0489