Refactor Armada Operator
We will move away from the usage of if/else block and will instead
make use of inheritance (armada base operator) and task specific
operators in our dags to execute the workflow.

We also added timeout option for client read so that the CI/CD team
can set other values for the execution of armada tasks (the default
time out is currently set to 1hr for all tasks).

Change-Id: I563fde76d91feae06a8a0298bc6eaf7cca1e66da