[Fix] Armada Operator/Dag - Task Id and Xcom Key
It was found that the airflow worker pod did not go through the
upgrade workflow even though the conditions for upgrade were met.
This happened after the armada operator was refactored [0]. This
patch set is meant to correct this behavior.

The CI/CD team informed us that we should increase the default time
out values for 'prepare_node_timeout' and 'post_apply_timeout' based
on their testing in the integration lab. Hence this patch set will also
update the default values for these 2 variables.

[0] https://review.gerrithub.io/#/c/404637/

Change-Id: Ic08ab04121b0bdae8673892cfd919475e812f98f