[ui] Add count of messages to commit configdocs CLI
The list of validations returned from commit configdocs
can be long and sometimes lose sight of errors. This change adds
a summary count of the severity level of messages returned at the
bottom of the output so as to clue the user into what might be closer
to the top.

Change-Id: I2a8921b22977b8ceab0afb1341b195c12dc49c67