Update execute_shipyard_action Script
The CI/CD team has pointed out that they are experiencing
problems in the pipeline while running the script as the
wrong position was queried for the action_lifecycle.

This has happened due to recent updates to the Shipyard API/CLI.

This p.s. will update the script, i.e. change from position
29 to 8.

Change-Id: Icd30fb43791e071207220f150886ae6201ed9e02