Clean up Operators that are not in use
The airflow_task_state Operator was created a while back before
the major revamp/refactoring of Shipyard/Airflow API/CLI and other
custom Operators. It is not used in any of the workflow and will
be removed.

The deckhand_get_design Operator is no longer needed after a new
logic to retrieve the last committed revision was introduced [0].
This task will now be performed during the initial stage of workflow
execution by Shipyard. Hence we will remove this Operator as well.

[0] https://review.gerrithub.io/#/c/att-comdev/shipyard/+/408457/

Change-Id: I6c0311d583b69056839c9e33e738f05f54a8288f