[FIX] Correct design ref use with Drydock client
Shipyard was passing a complete design reference dictionary to the
Drydock client, as it would for calling the Drydock API directly. This
causes an error to be reported (and subsequently ignored) during
validation of the DeploymentStrategy document. This change corrects the
behavior by passing only the href to the Drydock Client from the
deployment strategy validator

Change-Id: Idca0a69ec4ea11bf2bc4b520eb1512c0bdcd481b